Changed AbstractSocket implementation so that it will use non-blocking IO for send/receive and will only do a select() call once EAGAIN is returned.

+++ net/cpp/db/net/AbstractSocket.h
@@ -319,7 +319,7 @@ public:
     * 
     * @param timeout the send timeout in milliseconds.
     */
-   virtual void setSendTimeout(unsigned long timeout);
+   virtual void setSendTimeout(uint32_t timeout);
    
    /**
     * Gets the send timeout for this Socket. This is the amount of time that
@@ -327,7 +327,7 @@ public:
     * 
     * @return the send timeout in milliseconds.
     */
-   virtual unsigned long getSendTimeout();
+   virtual uint32_t getSendTimeout();
    
    /**
     * Sets the receive timeout for this Socket. This is the amount of time that
@@ -335,7 +335,7 @@ public:
     * 
     * @param timeout the receive timeout in milliseconds.
     */
-   virtual void setReceiveTimeout(unsigned long timeout);
+   virtual void setReceiveTimeout(uint32_t timeout);
    
    /**
     * Gets the receive timeout for this Socket. This is the amount of time that
@@ -343,7 +343,7 @@ public:
     * 
     * @return the receive timeout in milliseconds.
     */
-   virtual unsigned long getReceiveTimeout();
+   virtual uint32_t getReceiveTimeout();
    
    /**
     * Gets the number of Socket connections that can be kept backlogged while
