+++ net/cpp/db/net/SocketWrapper.h
@@ -211,7 +211,7 @@ public:
     * 
     * @param timeout the send timeout in milliseconds.
     */
-   virtual void setSendTimeout(unsigned long timeout);
+   virtual void setSendTimeout(uint32_t timeout);
    
    /**
     * Gets the send timeout for this Socket. This is the amount of time that
@@ -219,7 +219,7 @@ public:
     * 
     * @return the send timeout in milliseconds.
     */
-   virtual unsigned long getSendTimeout();
+   virtual uint32_t getSendTimeout();
    
    /**
     * Sets the receive timeout for this Socket. This is the amount of time that
@@ -227,7 +227,7 @@ public:
     * 
     * @param timeout the receive timeout in milliseconds.
     */
-   virtual void setReceiveTimeout(unsigned long timeout);
+   virtual void setReceiveTimeout(uint32_t timeout);
    
    /**
     * Gets the receive timeout for this Socket. This is the amount of time that
@@ -235,7 +235,7 @@ public:
     * 
     * @return the receive timeout in milliseconds.
     */
-   virtual unsigned long getReceiveTimeout();
+   virtual uint32_t getReceiveTimeout();
    
    /**
     * Gets the number of Socket connections that can be kept backlogged while
